# Improve the performance of the resource_logs endpoint